requirements.txt
selenium
kafka-python
nest_asyncio
beautifulsoup4